wt88047_64: Add GPS overlays from gps.conf
xtra server url need to be set, it is empty in framework, without those
xtra server set, it will not download the gps almanacs, ephemerides
data from wifi or radio network. It will need to be downloaded from
gps satellites directly which is very slow (> 10 minutes if download
from a single satellite).

[@ghostrider-reborn]:
* It seems Oreo gets these from overlays and not gps.conf anymore
* gps.conf was mainly read by QC's proprietary binaries, which were removed in N due to incompatibility.

* Values taken from wt88047's latest gps.conf (https://github.com/LineageOS/android_device_wingtech_wt88047/commit/a63520445ad1bc6a6c8fb074f5eaddde7bf618f1)

Co-authored-by: Ad!thya R <adithya.r02@outlook.com>
Signed-off-by: Cecell3000 <cecell3000@gmail.com>